+++ dm_collector_c/log_packet.h
@@ -1309,15 +1309,15 @@ const Fmt ModemDebug_Fmt [] = {
     {SKIP, NULL, 6},
 };
 
-bool is_log_packet (const char *b, int length);
-bool is_debug_packet (const char *b, int length);   //Yuanjie: test if it's a debugging message
+bool is_log_packet (const char *b, size_t length);
+bool is_debug_packet (const char *b, size_t length);   //Yuanjie: test if it's a debugging message
 
 // Given a binary string, try to decode it as a log packet.
 // Return a specailly formatted Python list that stores the decoding result.
 // If skip_decoding is True, only the header would be decoded.
-PyObject * decode_log_packet (const char *b, int length, bool skip_decoding);
+PyObject * decode_log_packet (const char *b, size_t length, bool skip_decoding);
 
 
-PyObject * decode_log_packet_modem (const char *b, int length, bool skip_decoding);
+PyObject * decode_log_packet_modem (const char *b, size_t length, bool skip_decoding);
 
 #endif  // __DM_COLLECTOR_C_LOG_PACKET_H__
